Add files via upload
Botium Toys: Audit scope and goals

An audit of Botium Toys’ cybersecurity program was performed. The audit needed to align with current business practices with industry standards and best practices.

What are the biggest risk to the organization? --- a. Botium Toys does not have the proper controls. --- b. Not compliant to the U.S. and international regulations and standards
Which controls are most essential to implement immediately versus in the future? --- a. Immediately: Administrative & Technical --- b. Future: Physical
Which compliance regulations does Botium Toys need to adhere to, to ensure the company keeps customer and vendor data safe, avoids fine, etc? --- a. GDPR, PCI DSS, SOC1 & SOC2